refactor: 패키지명 변경 common.entity -> common.domain
