+++ src/main/java/com/mojh/dailybudget/common/domain/BaseCreatedTimeEntity.java
@@ -1,9 +1,10 @@
-package com.mojh.dailybudget.common.entity;
+package com.mojh.dailybudget.common.domain;
 
 import lombok.Getter;
-import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
+import javax.persistence.Column;
 import javax.persistence.EntityListeners;
 import javax.persistence.MappedSuperclass;
 import java.time.LocalDateTime;
@@ -11,9 +12,9 @@ import java.time.LocalDateTime;
 @Getter
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-public abstract class BaseTimeEntity extends BaseCreatedTimeEntity {
+public abstract class BaseCreatedTimeEntity {
 
-    @LastModifiedDate
-    protected LocalDateTime updatedAt;
+    @CreatedDate
+    @Column(updatable = false)
+    protected LocalDateTime createdAt;
 }
-
