+++ src/main/java/com/mojh/dailybudget/common/domain/BaseTimeEntity.java
@@ -1,10 +1,9 @@
-package com.mojh.dailybudget.common.entity;
+package com.mojh.dailybudget.common.domain;
 
 import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import javax.persistence.Column;
 import javax.persistence.EntityListeners;
 import javax.persistence.MappedSuperclass;
 import java.time.LocalDateTime;
@@ -12,9 +11,9 @@ import java.time.LocalDateTime;
 @Getter
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-public abstract class BaseCreatedTimeEntity {
+public abstract class BaseTimeEntity extends BaseCreatedTimeEntity {
 
-    @CreatedDate
-    @Column(updatable = false)
-    protected LocalDateTime createdAt;
+    @LastModifiedDate
+    protected LocalDateTime updatedAt;
 }
+
